requirements.txt
Django==5.2.3
djangorestframework
django-cors-headers
channels
python-dotenv
geopy   